Fix #1826:  show subject interests on the profile page properly in Firefox (#1847)
* Fixed #1826, Interests now show properly in Firefox

Firefox's implementation of css doesn't collapse line feed for the nowrap option of white space.  I put the span for each interest on one line in the html file, and that fixed the problem.  This issue was reported to Mozilla in 2013, in bug 922773.

* Added a comment to explain style change

I added a comment in profile.html to explain why the span needs to be one line.

* Addressing another reviewer comment

Indented the new comment in html with the proper line of code.