Fixes part of #6240: Fix Flaky Exploration Card Finding (#8659)
* Protractor: Fix Flaky Exploration Card Finding

Finding exploration cards on the creator dashboard page was flaky
because sometimes the protractor test would search for the cards before
the cards had loaded. Further, the logic used to identify matching cards
failed when the card contained text besides the exploration title. This
commit resolves these issues by:

* Waiting for an exploration card to appear before searching for them
* When filtering cards by title, consider cards to match if their text
  equals the target title or begins with the target title concatenated
  with a newline. This works because when other text is included in the
  card, a newline separates that text from the title.

* Style fixes

* Replace presenceOf with visibilityOf

We were waiting for a card to exist in the DOM (whether or not it was
visible), but we really do want it to be visible before clicking it.
Further, by relying on `visibilityOf`, we can use existing code. This
commit makes this switch.