Fix #10247: Set CodeMirror viewportMargin to 100 and fix e2e flake (#10248)
* Set CodeMirror viewportMargin to 100

CodeMirror uses the viewportMargin property to determine how many
lines are shown above and below the viewport currently scrolled
into view. The number of lines shown depends on the viewPort height
this may vary depending on the screensize and the zoom level set
in the browser. In a deskstop screen with a 1920x1080 resolution,
the History tab diff (this is where CodeMirror is used), shows
approximately 20 lines (and renders 20 in the top and 20 in
the bottom that are hidden from view). Scrolling upwards or
downwards (E.g. scrolling 40+ lines) quickly causes the rendering
to be a bit slow because those lines are not yet processed by
CodeMirror. By increasing the viewportMargin to 100, we allow more
lines to be processed before scrolling. This make the scrolling
experience a bit better (E.g. scrolling past 50-100 lines of code
does not need any processing). This also helps protractor tests run
smoothly because it does not have to "figure out" if the lines
on the screen have been processed or not. A commonly occuring e2e
flake (#10247) should be fixed by this change.

* tweak scroll logic

* fix lint error

* rewrite logic for comparing history diff in the e2e test

* fix lint issues