Fix #11023: Improve robustness of platform parameter evaluation (#11074)
* First attempt to improve platform param robustness.

This change allows None client type & user locales rather than producing
validation errors, and instead leads to platform parameters evaluating
to default in this case.

However, it seems that server evaluation currently ignores both of these
types today which seems incorrect. This was missed in the initial
implementation and more thought is needed.

* Mostly fix server context init for platform params.

This properly defines the user locale and client type with parity with
the web frontend, though it introduces more conflation with both terms.
It also doesn't define a valid version for the backend yet. The latter
changes will happen in a follow-up issue (#11073).

* Add tests.

Added more tests to clarify the new server values being specified for
platform parameter evaluation.

* Address lint error (for existing code).

* Post-merge fix.