Fix 17072 : Make submit question checkbox accessible (#18037)
Make the checkbox that marks an answer as correct on the submit question page accessible by a screen reader. This is done by adding a label tag to the existing description and binding it to the check box.The fix is made on 2 pages, the modal dialog and the page behind.

Co-authored-by: Praneeth <praneethg2001@gmail.com>