Upgrade PyPI development dependencies (#18165)
* Upgrade PyPI Development Dependencies

Changelogs for the upgraded dependencies in `requirements_dev.in` were
skimmed. After this commit is merged, developers should run `python -m
scripts.install_python_dev_dependencies` to make sure their
installations are up-to-date.

* Upgrade a Few PyPI Prod Dependencies

Upgrade only the PyPI prod dependencies that were also upgraded as part
of the dev dependencies upgrade. After this commit is merged, developers
should run `python -m scripts.install_third_party` to make sure they
have the updated dependencies.

* Remove obsolete TODO

* Update pylint_extensions.py for latest pylint

* Update pylint config for new pylint version

* Fix docstring validity checks and linting

* Allow assigning lambdas to variables

* Fix docstrings_checker.py tests

* Fix lint errors

* Upgrade pip

* Reinstall installation tools during build

* Add core.tests.test_utils_test to coverage ignore

Before the dependency upgrade, coverage reports from backend test runs
on develop show that core.tests.test_utils_test was not being executed
for some reason. With the upgrade, it is now being included in the test
suite, but since it does not have complete per-file coverage, it needs
to be added to the ignore list.

* Fix backend tests

* Fix protoc version

* Revert protobuf upgrade

* Fix build.py backend test

* Fix Coverage by Removing Unnecessary if

Protobuf is installed in all environments now as part of our development
tools, so there's no need to check whether a google package is in
Python's module path.

* Increase core/tests/test_utils.py coverage

* Remove unused call_through arg to swap_with_call_counter

* Add backend tests for core/test/test_utils.py

* Remove unused GenericTestBase creation funcs

* Remove unused imports

* Add types to mock function

* Remove unused test_utils.py code

* Add test to use unused test_utils.py functionality

* Fix coverage errors

* Address review comments

* Fix typo

* Fix backend test