Fix #18654: Migrates admin page to angular (#18855)
* Move admin page to angular

* Add newline at the EOF

* Move gaurd to module route

* Add gaurd file

* Misc lint fixes

* Remove downgrade from child components

* Render a 404 page from the admin page

A route guard works best when we redirect users to a different page
if a condition does not pass. 404 is not a registered page in our app
so in this case, we can just render the page from the admin root page
instead. oppia-blog-author-profile-page-root also handles it in a
similar way.

A new module for 404 was created because the existing 404 module also
adds an empty path route (''). This conflicts with other empty path
routes in modules. So the new module imports the existing (non-route)
module and just adds a route to it. The routed version is used by the
app routing modules while the non routed version is used everywhere
else.

* Make the migrated admin page similar to the earlier one

Remove the oppia footer and change title to match the new page.

* Remove the earlier admin page

* Remove previously commented code

* MyPy Fix: Match get signature to super class

* Nit

* Add codeowner for base component file

* Convert title and meta to abstract properties

BaseRootComponent will not compile unless it is decorated as an
angular component since it uses angular features. Once decorated,
depedency injection will throw an error because it cannot find
providers for title and meta.

* Add tests for BaseRootComponent

* Show 401 unauthorized when user is not super admin

* Redirect admin to login page for users not logged in

* Add IsLoggedInGuard to services index

* Add tests for admin page root component

* Fix eslint error

* Use angular guard to redirect users to 401 error page

Also host error pages on the root domain.

* Migrate error pages to angular router

* Rename ADMIN route and use ERROR route

* Remove error pages from webpack

* Revert "Remove error pages from webpack"

This reverts commit f086794ef1a3c7dff6b00d07a84c76480ebdeb64.

* Create two error page modules

One to handle error page routing from the backend and one that handles
error page routing from the angular router.

* Fix lint errors

* Fix TS compilation errors

* Remove admin rights test

These rights are checked on the frontend router now.

* Add router testing module to the services import

* Add an open access decorator to AdminPage

Page can be accessed by anyone. Permissioning is handled by the
angular router on the frontend.

* Add a test to check the response of /admin page

* Move common declarations in a common module

* Add AdminPageRootComponent tests

* Fix error module depedencies

* Comment changes

* Rename existing error pages to old

And rename error-page-migrated.module.ts to error-page.module.ts

* Use oppia root handler for the admin page

* Remove unused import

---------

Co-authored-by: Kapil Agarwal <kapilag96@gmail.com>